add AbstractGate class and refactor code
