+++ pyqtorch/modules/parametric.py
@@ -2,29 +2,25 @@ from __future__ import annotations
 
 import torch
 from numpy.typing import ArrayLike
-from torch.nn import Module
 
 from pyqtorch.core.batched_operation import (
     _apply_batch_gate,
     create_controlled_batch_from_operation,
 )
 from pyqtorch.core.utils import OPERATIONS_DICT
+from pyqtorch.modules.utils import AbstractGate, rot_matrices
 
 
-class RotationGate(Module):
+class RotationGate(AbstractGate):
     n_params = 1
 
     def __init__(self, gate: str, qubits: ArrayLike, n_qubits: int):
-        super().__init__()
-        self.qubits = qubits
-        self.n_qubits = n_qubits
+        super().__init__(qubits, n_qubits)
         self.gate = gate
         self.register_buffer("imat", OPERATIONS_DICT["I"])
         self.register_buffer("paulimat", OPERATIONS_DICT[gate])
 
     def matrices(self, thetas: torch.Tensor) -> torch.Tensor:
-        # NOTE: thetas are assumed to be of shape (1,batch_size) or (batch_size,) because we
-        # want to allow e.g. (3,batch_size) in the U gate.
         theta = thetas.squeeze(0) if thetas.ndim == 2 else thetas
         batch_size = len(theta)
         return rot_matrices(theta, self.paulimat, self.imat, batch_size)
@@ -38,28 +34,10 @@ class RotationGate(Module):
         return self.apply(mats, state)
 
     def extra_repr(self) -> str:
-        return f"qubits={self.qubits}, n_qubits={self.n_qubits}"
+        return super().extra_repr()
 
 
-def rot_matrices(
-    theta: torch.Tensor, P: torch.Tensor, I: torch.Tensor, batch_size: int  # noqa: E741
-) -> torch.Tensor:
-    """
-    Returns:
-        torch.Tensor: a batch of gates after applying theta
-    """
-    cos_t = torch.cos(theta / 2).unsqueeze(0).unsqueeze(1)
-    cos_t = cos_t.repeat((2, 2, 1))
-    sin_t = torch.sin(theta / 2).unsqueeze(0).unsqueeze(1)
-    sin_t = sin_t.repeat((2, 2, 1))
-
-    batch_imat = I.unsqueeze(2).repeat(1, 1, batch_size)
-    batch_operation_mat = P.unsqueeze(2).repeat(1, 1, batch_size)
-
-    return cos_t * batch_imat - 1j * sin_t * batch_operation_mat
-
-
-class U(Module):
+class U(AbstractGate):
     """Parametrized arbitrary rotation along the axes of the Bloch sphere
 
     The angles `phi, theta, omega` in tensor format, applied as:
@@ -70,9 +48,7 @@ class U(Module):
     n_params = 3
 
     def __init__(self, qubits: ArrayLike, n_qubits: int):
-        super().__init__()
-        self.qubits = qubits
-        self.n_qubits = n_qubits
+        super().__init__(qubits, n_qubits)
 
         self.register_buffer("a", torch.tensor([[1, 0], [0, 0]], dtype=torch.cdouble).unsqueeze(2))
         self.register_buffer("b", torch.tensor([[0, 1], [0, 0]], dtype=torch.cdouble).unsqueeze(2))
@@ -114,16 +90,14 @@ class U(Module):
         return self.apply(mats, state)
 
     def extra_repr(self) -> str:
-        return f"qubits={self.qubits}, n_qubits={self.n_qubits}"
+        return super().extra_repr()
 
 
-class ControlledRotationGate(Module):
+class ControlledRotationGate(AbstractGate):
     n_params = 1
 
     def __init__(self, gate: str, qubits: ArrayLike, n_qubits: int):
-        super().__init__()
-        self.qubits = qubits
-        self.n_qubits = n_qubits
+        super().__init__(qubits, n_qubits)
         self.gate = gate
         self.register_buffer("imat", OPERATIONS_DICT["I"])
         self.register_buffer("paulimat", OPERATIONS_DICT[gate])
@@ -143,7 +117,7 @@ class ControlledRotationGate(Module):
         return self.apply(mats, state)
 
     def extra_repr(self) -> str:
-        return f"qubits={self.qubits}, n_qubits={self.n_qubits}"
+        return super().extra_repr()
 
 
 class RX(RotationGate):
@@ -176,18 +150,14 @@ class CRZ(ControlledRotationGate):
         super().__init__("Z", qubits, n_qubits)
 
 
-class CPHASE(Module):
+class CPHASE(AbstractGate):
     n_params = 1
 
     def __init__(self, qubits: ArrayLike, n_qubits: int):
-        super().__init__()
-        self.qubits = qubits
-        self.n_qubits = n_qubits
+        super().__init__(qubits, n_qubits)
         self.register_buffer("imat", torch.eye(4, dtype=torch.cdouble))
 
     def matrices(self, thetas: torch.Tensor) -> torch.Tensor:
-        # NOTE: thetas are assumed to be of shape (1,batch_size) or (batch_size,) because we
-        # want to allow e.g. (3,batch_size) in the U gate.
         theta = thetas.squeeze(0) if thetas.ndim == 2 else thetas
         batch_size = len(theta)
         mat = self.imat.repeat((batch_size, 1, 1))
@@ -205,4 +175,4 @@ class CPHASE(Module):
         return self.apply(mats, state)
 
     def extra_repr(self) -> str:
-        return f"qubits={self.qubits}, n_qubits={self.n_qubits}"
+        return super().extra_repr()
